Backend: make vehicles change src and dst upon first boarding in garage
Problem Solved
Vehicles used to keep the default slices for their source and destination.

Rationale
This led to inconsistencies with respect to the source and destination of a traveller who boards them
while they are still parked in a garage, causing a wrong route computation for the vehicle and,
therefore, faulty behaviours at runtime.